refactor: Remove App and use page Home
