--- src/components/App/index.js
@@ -1,1 +1,0 @@
-export { default } from './App.component';
