+++ src/components/Card/Card.component.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function App() {
+  return (
+    <div className="card">
+      <img src="" alt="" />
+      <div className="card__content" />
+    </div>
+  );
+}
+
+export default App;
