+++ src/components/Card/Navbar.styles.css
@@ -1,0 +1,14 @@
+.card {
+  transition: transform 100ms;
+  display: flex;
+  margin: 10px;
+  width: 550px;
+  height: 170px;
+  margin: 10px;
+  padding: 10px;
+  border-radius: 5px;
+}
+
+.card:hover {
+  transform: scale(1.05);
+}
