--- src/pages/Login/Login.styles.css
@@ -1,47 +1,0 @@
-.login {
-  width: 300px;
-}
-
-.login h1 {
-  text-align: center;
-  letter-spacing: -1px;
-}
-
-.login-form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.form-group {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 1rem;
-}
-
-.form-group strong {
-  display: block;
-  font-weight: 700;
-  text-transform: capitalize;
-  margin-bottom: 0.4rem;
-}
-
-.form-group input {
-  color: white;
-  font-size: 1.2rem;
-  width: 100%;
-  padding: 0.4rem 0.6rem;
-  border-radius: 3px;
-  border: 1px solid white;
-  background-color: rgba(0, 0, 0, 0.1);
-}
-
-.login-form button[type='submit'] {
-  width: 5rem;
-  margin-top: 1rem;
-  padding: 0.4rem 0.6rem;
-  font-size: 1.2rem;
-  border: none;
-  border-radius: 3px;
-}
