--- src/pages/NotFound/index.js
@@ -1,1 +1,0 @@
-export { default } from './NotFound.page';
